Moved viewer thread func into own file
