+++ src/main.cpp
@@ -31,6 +31,7 @@
 #include "input.hpp"
 #include "util.hpp"
 
+#include "viewer.hpp"
 #include "map_editor.hpp"
 
 #include "editor_events.hpp"
@@ -41,156 +42,6 @@ constexpr int 	CONSOLE_HEIGHT		= 256;
 constexpr char 	DISPLAY_TITLE[]		= "Axe DnD Map";
 
 using std_clk = std::chrono::steady_clock;
-
-struct ThreadArgs
-{
-	std::string bmp_path;
-	ALLEGRO_MUTEX *mutex;
-	ALLEGRO_COND *cond;
-	ALLEGRO_EVENT_SOURCE *event_source;
-};
-
-ALLEGRO_DISPLAY *createDisplay(const char* title, int width, int height, int flags)
-{
-	ALLEGRO_DISPLAY *d = nullptr;
-
-	al_set_new_display_flags(flags);
-
-	std::string full_title = DISPLAY_TITLE;
-	full_title += " - ";
-	full_title += title;
-	al_set_new_window_title(full_title.c_str());
-
-	d = al_create_display(width, height);
-
-	if (!d)
-	{
-		std::cerr << "Failed to create display!" << std::endl;
-		exit(EXIT_FAILURE);
-	}
-
-	return d;
-}
-
-static void *thread_func(ALLEGRO_THREAD* thr, void* arg)
-{
-	ALLEGRO_DISPLAY* display = nullptr;
-	ALLEGRO_EVENT_QUEUE* evq = nullptr;
-	ALLEGRO_TIMER* timer = nullptr;
-	ALLEGRO_EVENT ev;
-	std_clk::time_point current_time;
-	double delta_time;
-	bool running = true;
-	bool redraw = true;
-
-	ThreadArgs *thargs = (ThreadArgs*)arg;
-	View view;
-	bool lerping = false;
-	constexpr double lerp_time = 1.5;
-	double elapsed = 0.0;
-	vec2d view_start;
-	vec2d view_target;
-	ALLEGRO_BITMAP* bmp;
-
-	display = createDisplay("Viewer", DEFAULT_WIND_WIDTH, DEFAULT_WIND_HEIGHT - CONSOLE_HEIGHT, ALLEGRO_RESIZABLE | ALLEGRO_WINDOWED);
-	timer = al_create_timer(1.0 / 60.0);
-	evq = al_create_event_queue();
-
-	al_register_event_source(evq, al_get_timer_event_source(timer));
-	al_register_event_source(evq, al_get_display_event_source(display));
-	al_register_event_source(evq, thargs->event_source);
-
-	bmp = al_load_bitmap(thargs->bmp_path.c_str());
-	if (!bmp)
-	{
-		std::cerr << "Failed to load bitmap in thread" << std::endl;
-		if (display) al_destroy_display(display);
-		if (timer) al_destroy_timer(timer);
-		if (evq) al_destroy_event_queue(evq);
-		return NULL;
-	}
-
-	al_start_timer(timer);
-	auto last_time = std_clk::now();
-	while (running)
-	{
-		if (al_get_thread_should_stop(thr))
-		{
-			break;
-		}
-
-		al_wait_for_event(evq, &ev);
-	
-		vec2d diff;
-		switch (ev.type)
-		{
-			case AXE_EDITOR_EVENT_MOVE_VIEW:
-				view_target.x = static_cast<double>(ev.user.data1);
-				view_target.y = static_cast<double>(ev.user.data2);
-				view_start = view.world_pos;
-				lerping = true;
-				elapsed = 0.0;
-			break;
-
-			case AXE_EDITOR_EVENT_SCALE_VIEW:
-				std::cout << "AXE_EDITOR_EVENT_SCALE_VIEW\n" << "\tev.user.data1 = " << ev.user.data1 << "\n";
-				if (ev.user.data1 > 0)
-				{
-					view.scale += { 0.1, 0.1 };
-				}
-				else if (ev.user.data1 < 0)
-				{
-					view.scale -= { 0.1, 0.1 };
-				}
-			break;
-
-			case ALLEGRO_EVENT_TIMER:
-				current_time = std_clk::now();
-				delta_time = std::chrono::duration<double>(current_time - last_time).count();
-				last_time = current_time;
-				elapsed += delta_time;
-				redraw = true;
-
-				if (lerping && elapsed <= lerp_time)
-				{
-					view.world_pos = vec_lerp(view_start, view_target, easeInAndOutQuart(elapsed / lerp_time));
-				}
-				else lerping = false;
-			break;
-
-			case ALLEGRO_EVENT_DISPLAY_CLOSE:
-				running = false;
-			break;
-
-			case ALLEGRO_EVENT_DISPLAY_RESIZE:
-				al_acknowledge_resize(display);
-			break;
-
-			default:
-			break;
-		}
-
-		if (al_event_queue_is_empty(evq) && redraw)
-		{
-			al_clear_to_color(al_map_rgb(45, 45, 45));
-
-			int w = al_get_display_width(display);
-			int h = al_get_display_height(display);
-
-			al_draw_bitmap_region(bmp, view.world_pos.x - (w/2),view.world_pos.y - (h/2), w, h, 0, 0, 0);
-
-			al_flip_display();
-			redraw = false;
-		}
-	}
-
-	al_destroy_bitmap(bmp);
-	al_destroy_timer(timer);
-	al_destroy_event_queue(evq);
-	al_destroy_display(display);
-
-	return NULL;
-}
 
 int main(int argc, char ** argv)
 {
@@ -216,7 +67,7 @@ int main(int argc, char ** argv)
 		exit(EXIT_FAILURE);
 	}
 
-	main_display = createDisplay("Editor", DEFAULT_WIND_WIDTH, DEFAULT_WIND_HEIGHT, ALLEGRO_WINDOWED | ALLEGRO_RESIZABLE);
+	main_display = createDisplay(std::string(DISPLAY_TITLE) + " - Editor", DEFAULT_WIND_WIDTH, DEFAULT_WIND_HEIGHT, ALLEGRO_WINDOWED | ALLEGRO_RESIZABLE);
 
 	al_init_image_addon();
 	al_init_font_addon();
@@ -240,6 +91,8 @@ int main(int argc, char ** argv)
 	al_init_user_event_source(&editor_event_source);
 
 	thargs.event_source = &editor_event_source;
+	thargs.display_title = std::string(DISPLAY_TITLE) + " - Viewer";
+	thargs.display_size = { DEFAULT_WIND_WIDTH, DEFAULT_WIND_HEIGHT };
 
 	MapEditor map_editor(m_input, editor_event_source, {0, 0}, { getScreenSize().x, getScreenSize().y - CONSOLE_HEIGHT });
